fix cast aliquota quote e salesorder

+++ RSMNG.TAUMEDIKA.Plugins/Quote/PreUpdate.cs
@@ -54,7 +54,6 @@ namespace RSMNG.TAUMEDIKA.Plugins.Quote
             target.AddWithRemove(quote.res_recipient, destination);
             #endregion
 
-            
             #region Valorizzazione automatica del campo Motivo Stato Precedente
             PluginRegion = "Valorizzazione automatica del campo Motivo Stato Precedente";
 
@@ -116,9 +115,9 @@ namespace RSMNG.TAUMEDIKA.Plugins.Quote
                 {
                     Entity enQuote = quoteCollection.Entities[0];
 
-                    decimal importoSpesaAccessoria = enQuote.ContainsAttributeNotNull(quote.freightamount) ?  enQuote.GetAttributeValue<Money>(quote.freightamount).Value : 0;
-                    decimal aliquota =  enQuote.ContainsAliasNotNull($"{res_vatnumber.logicalName}.{res_vatnumber.res_rate}") ?  enQuote.GetAliasedValue<Money>($"{res_vatnumber.logicalName}.{res_vatnumber.res_rate}").Value : 1;
-                    decimal aliquotaImportoSpesaAccessoria = importoSpesaAccessoria * ((aliquota == 0 ? 1:aliquota) / 100);
+                    decimal importoSpesaAccessoria = enQuote.ContainsAttributeNotNull(quote.freightamount) ? enQuote.GetAttributeValue<Money>(quote.freightamount).Value : 0;
+                    decimal aliquota = enQuote.ContainsAliasNotNull($"{res_vatnumber.logicalName}.{res_vatnumber.res_rate}") ? (decimal)enQuote.GetAliasedValue<Money>($"{res_vatnumber.logicalName}.{res_vatnumber.res_rate}").Value : 1;
+                    decimal aliquotaImportoSpesaAccessoria = importoSpesaAccessoria * ((aliquota == 0 ? 1 : aliquota) / 100);
 
                     totaleIva += aliquotaImportoSpesaAccessoria;
                     totaleImponibile = totaleProdotti + importoSpesaAccessoria;
